+++ lib/src/core/alphabet.dart
@@ -1,15 +1,13 @@
 // Copyright (c) 2023, Sudipto Chandra
 // All rights reserved. Check LICENSE file for details.
 
+import 'dart:typed_data';
+
+import 'byte.dart';
 import 'decoder.dart';
-import 'encoder.dart';
 
-class AlphabetEncoder extends BitEncoder {
-  final int bits;
+class AlphabetEncoder extends ByteEncoder {
   final List<int> alphabet;
-
-  @override
-  final int source = 8;
 
   /// The padding character.
   ///
@@ -25,22 +23,36 @@ class AlphabetEncoder extends BitEncoder {
   /// - The output array will be padded with the [padding] to make the length of
   ///   the array to be divisible by [source].
   const AlphabetEncoder({
-    required this.bits,
+    required int bits,
     required this.alphabet,
     this.padding,
-  });
+  }) : super(bits: bits);
 
   @override
   int get target => bits;
 
   @override
-  Iterable<int> convert(Iterable<int> input) {
-    var out = super.convert(input).map((x) => alphabet[x]).toList();
-    int l = out.length * target;
-    if (padding != null) {
-      for (; (l & 7) != 0; l += target) {
-        out.add(padding!);
-      }
+  Uint8List convert(List<int> input) {
+    int i, n;
+    var encoded = super.convert(input) as Uint8List;
+    n = encoded.length;
+    for (i = 0; i < n; ++i) {
+      encoded[i] = alphabet[encoded[i]];
+    }
+    if (padding == null) {
+      return encoded;
+    }
+
+    n = encoded.length;
+    for (i = n * target; (i & 7) != 0; i += target) {
+      n++;
+    }
+    var out = Uint8List(n);
+    for (i = 0; i < encoded.length; ++i) {
+      out[i] = encoded[i];
+    }
+    for (; i < n; ++i) {
+      out[i] = padding!;
     }
     return out;
   }
@@ -75,7 +87,7 @@ class AlphabetDecoder extends BitDecoder {
   int get source => bits;
 
   @override
-  Iterable<int> convert(Iterable<int> encoded) {
+  List<int> convert(List<int> encoded) {
     int x;
     return super.convert(encoded.map((y) {
       if (y == padding) return -1;
@@ -83,6 +95,6 @@ class AlphabetDecoder extends BitDecoder {
         throw FormatException('Invalid character $y');
       }
       return x;
-    }));
+    }).toList());
   }
 }