+++ test/base8_test.dart
@@ -8,55 +8,87 @@ import 'utils.dart';
 
 void main() {
   group('Test base8', () {
-    group("encoding", () {
-      test('[] => empty string', () {
-        expect(toOctal([]), "");
+    group('[] <=> empty string', () {
+      final input = <int>[];
+      final output = "";
+      test('encoding', () {
+        expect(toOctal(input), output);
       });
-      test('[1] => 002', () {
-        expect(toOctal([1]), "002");
-      });
-      test('[0, 1] => 000004', () {
-        expect(toOctal([0, 1]), "000004");
-      });
-      test('[1, 0] => 002000', () {
-        expect(toOctal([1, 0]), "002000");
-      });
-      test('[7] => 016', () {
-        expect(toOctal([7]), "016");
-      });
-      test('[10] => 024', () {
-        expect(toOctal([10]), "024");
-      });
-      test('[0, 10] => 000050', () {
-        expect(toOctal([0, 10]), "000050");
+      test('decoding', () {
+        expect(fromOctal(output), input);
       });
     });
-    group("decoding", () {
-      test('empty string => []', () {
-        expect(fromOctal(""), []);
+    group('[1] <=> 001', () {
+      final input = <int>[1];
+      final output = "001";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
       });
-      test('12 |001 010| => invalid length', () {
-        expect(() => fromOctal("12"), throwsFormatException);
-      });
-      test('012 |000 001 01|0| => [5]', () {
-        expect(fromOctal("012"), equals([5]));
-      });
-      test('0012 |000 000 00|1 010| => invalid length', () {
-        expect(() => fromOctal("0012"), throwsFormatException);
-      });
-      test('00012 |000 000 00|0 001 010| => invalid length', () {
-        expect(() => fromOctal("00012"), throwsFormatException);
-      });
-      test('000012 |000 000 00|0 000 001 0|10| => invalid length', () {
-        expect(() => fromOctal("000012"), throwsFormatException);
-      });
-      test('0000012 |000 000 00|0 000 000 0|01 010| => : invalid length', () {
-        expect(() => fromOctal("0000012"), throwsFormatException);
-      });
-      test('00000012 |000 000 00|0 000 000 0|00 001 010| => [0, 0, 10]', () {
-        expect(fromOctal("00000012"), equals([0, 0, 10]));
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
       });
     });
+    group('[0, 1] <=> 000001', () {
+      final input = <int>[0, 1];
+      final output = "000001";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+    group('[1, 0] <=> 000400', () {
+      final input = <int>[1, 0];
+      final output = "000400";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+    group('[7] <=> 007', () {
+      final input = <int>[7];
+      final output = "007";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+    group('[10] <=> 012', () {
+      final input = <int>[10];
+      final output = "012";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+    group('[0, 10] <=> 000012', () {
+      final input = <int>[0, 10];
+      final output = "000012";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+    group('[1, 2, 3, 4, 5, 6, 7, 8] => 0004020060200501403410', () {
+      var input = [1, 2, 3, 4, 5, 6, 7, 8];
+      final output = "0004020060200501403410";
+      test('encoding', () {
+        expect(toOctal(input), equals(output));
+      });
+      test('decoding', () {
+        expect(fromOctal(output), equals(input));
+      });
+    });
+
     test('encoding <-> decoding', () {
       for (int i = 0; i < 100; ++i) {
         var b = randomBytes(i);
@@ -64,9 +96,18 @@ void main() {
         expect(fromOctal(r), equals(b), reason: 'length $i');
       }
     });
-    group('decoding with invalid chars', () {
-      test('0182', () {
-        expect(() => fromOctal("0182"), throwsFormatException);
+
+    group('decoding edge cases', () {
+      test('partial message', () {
+        expect(fromOctal("12"), equals([10]));
+      });
+      test('long partial message', () {
+        final input = '4020060200501403410';
+        final output = [1, 2, 3, 4, 5, 6, 7, 8];
+        expect(fromOctal(input), equals(output));
+      });
+      test('182', () {
+        expect(() => fromOctal("182"), throwsFormatException);
       });
       test('-10', () {
         expect(() => fromOctal("-10"), throwsFormatException);