+++ test/bigint_test.dart
@@ -69,6 +69,46 @@ void main() {
         expect(out2, equals(inp), reason: 'length $i');
       }
     });
+    test('encoding 32-bit big endian in MSB first order', () {
+      var inp = <int>[1, 2, 3, 4];
+      var out = BigInt.from(0x01020304);
+      expect(toBigInt(inp, msbFirst: true), equals(out));
+    });
+    test('encoding 32-bit big endian in LSB first order', () {
+      var inp = <int>[4, 3, 2, 1];
+      var out = BigInt.from(0x01020304);
+      expect(toBigInt(inp, msbFirst: false), equals(out));
+    });
+    test('decoding 32-bit big endian in MSB first order', () {
+      var inp = <int>[1, 2, 3, 4];
+      var out = BigInt.from(0x01020304);
+      expect(fromBigInt(out, msbFirst: true), equals(inp));
+    });
+    test('decoding 32-bit big endian in LSB first order', () {
+      var inp = <int>[4, 3, 2, 1];
+      var out = BigInt.from(0x01020304);
+      expect(fromBigInt(out, msbFirst: false), equals(inp));
+    });
+    test('encoding 64-bit big endian in MSB first order', () {
+      var inp = <int>[1, 2, 3, 4, 5, 6, 7, 8];
+      var out = BigInt.from((0x01020304 << 32) | 0x05060708);
+      expect(toBigInt(inp, msbFirst: true), equals(out));
+    }, tags: ['vm-only']);
+    test('encoding 64-bit big endian in LSB first order', () {
+      var inp = <int>[8, 7, 6, 5, 4, 3, 2, 1];
+      var out = BigInt.from((0x01020304 << 32) | 0x05060708);
+      expect(toBigInt(inp, msbFirst: false), equals(out));
+    }, tags: ['vm-only']);
+    test('decoding 64-bit big endian in MSB first order', () {
+      var inp = <int>[1, 2, 3, 4, 5, 6, 7, 8];
+      var out = BigInt.from((0x01020304 << 32) | 0x05060708);
+      expect(fromBigInt(out, msbFirst: true), equals(inp));
+    }, tags: ['vm-only']);
+    test('decoding 64-bit big endian in LSB first order', () {
+      var inp = <int>[8, 7, 6, 5, 4, 3, 2, 1];
+      var out = BigInt.from((0x01020304 << 32) | 0x05060708);
+      expect(fromBigInt(out, msbFirst: false), equals(inp));
+    }, tags: ['vm-only']);
     test('big-endian encoding <-> decoding', () {
       for (int i = 0; i < 100; ++i) {
         var inp = [1, ...randomBytes(i)];